backend/requirements.txt
annotated-types
anyio
cachetools
certifi
charset-normalizer
click
dnspython
email_validator
fastapi
fastapi-cli
filelock
fsspec
google-ai-generativelanguage
google-api-core
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpciorc1
grpcio-status
h11
httpcore
httplib2
httptools
httpx
huggingface-hub
idna
Jinja2
markdown-it-py
MarkupSafe
mdurl
mpmath
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
packaging
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
Pygments
pyparsing
python-dotenv
python-multipart
PyYAML
regex
requests
rich
rich-toolkit
rsa
safetensors
setuptools
shellingham
sniffio
starlette
sympy
tokenizers
torch
tqdm
transformers
triton
typer
typing-inspection
typing_extensions
uritemplate
urllib3
uvicorn
uvloop
watchfiles
websockets
